Add gpio implementation; Add basic example;

+++ ext/bonekit/pin.h
@@ -11,7 +11,7 @@ extern "C" {
 struct pin_s
 {
   unsigned _gpio;
-  uint8_t _mode;
+  int _mode;
   int _value;
 };
 
@@ -19,8 +19,8 @@ typedef struct pin_s pin_t;
 
 pin_t * pin_create(unsigned);
 void pin_destroy(pin_t *);
-uint8_t pin_mode(pin_t *);
-void pin_set_mode(pin_t *, uint8_t);
+int pin_mode(pin_t *);
+void pin_set_mode(pin_t *, int);
 int pin_value(pin_t *);
 void pin_set_value(pin_t *, int);
